fix: move sizeBehavior
